Fix a heap buffer underread in set_buf_break()
If an opening parenthesis follows a comment with a text, a read from
an invalid address happens in set_buf_break():

    $ printf '/*a*/()' | valgrind -- ./src/indent - -o /dev/null
    ==28887== Memcheck, a memory error detector
    ==28887== Copyright (C) 2002-2022, and GNU GPL'd, by Julian Seward et al.
    ==28887== Using Valgrind-3.22.0 and LibVEX; rerun with -h for copyright info
    ==28887== Command: ./src/indent - -o /dev/null
    ==28887==
    ==28887== Invalid read of size 2
    ==28887==    at 0x409989: set_buf_break (output.c:319)
    ==28887==    by 0x401FE7: indent_main_loop (indent.c:640)
    ==28887==    by 0x4022A7: indent (indent.c:759)
    ==28887==    by 0x40294E: indent_single_file (indent.c:1004)
    ==28887==    by 0x402A1C: indent_all (indent.c:1042)
    ==28887==    by 0x402BD0: main (indent.c:1123)
    ==28887==  Address 0x4a5facc is 4 bytes before a block of size 16 alloc'd
    ==28887==    at 0x4849E60: calloc (vg_replace_malloc.c:1595)
    ==28887==    by 0x408B61: xmalloc (globs.c:42)
    ==28887==    by 0x40765E: init_parser (parse.c:73)
    ==28887==    by 0x402B1F: main (indent.c:1101)

It happens when checking an indentation level of the outer scope by indexing
parser_state_tos->paren_indents[]:

    level = parser_state_tos->p_l_follow;
    [...]
    /* Did we just parse a bracket that will be put on the next line
     * by this line break? */
    if ((*token == '(') || (*token == '['))
        --level;    /* then don't take it into account */
    [...]
    if (level == 0) {
    } else {
→       if (parser_state_tos->paren_indents[level - 1] < 0) {...}
    }

The cause is a special case for moving opening parentheses and
brackets to a next line. If parser_state_tos->p_l_follow is zero
(like in the reproducer), the index evaluates to -2 and goes out of
range of the paren_indents array.

This patch simply prevents from decreasing the index under zero when
formating the code. Maybe it leaves some piece of code unformated, but
it's safe.

I checked all places where p_l_follow is set (it is only in
handletoken.c) and they corretly prevent from decrasing it under
zero. That keeps set_buf_break() in output.c as the culprit.

<https://lists.gnu.org/archive/html/bug-indent/2024-01/msg00000.html>

Signed-off-by: Petr Písař <ppisar@redhat.com>